app/requirements.txt
fastapi
requests
sqlalchemy
uvicorn
psycopg2-binary
python-dotenv
